+++ src/main/java/domain/card/CardMachine.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import constant.PolicyConstant;
 import constant.Rank;
