+++ src/main/java/domain/card/Hand.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import constant.PolicyConstant;
 import java.util.ArrayList;
