+++ src/main/java/domain/participant/Participant.java
@@ -1,8 +1,8 @@
 package domain.participant;
 
 import constant.PolicyConstant;
-import domain.Card;
-import domain.Hand;
+import domain.card.Card;
+import domain.card.Hand;
 import java.util.List;
 
 public class Participant {
