+++ src/main/java/service/BlackjackService.java
@@ -3,8 +3,8 @@ package service;
 import constant.PlayerAction;
 import constant.PolicyConstant;
 import constant.Result;
-import domain.CardMachine;
 import domain.bet.Money;
+import domain.card.CardMachine;
 import domain.participant.Dealer;
 import domain.participant.Participant;
 import domain.participant.Player;
