+++ src/test/java/domain/DealerTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import constant.Rank;
 import constant.Suit;
+import domain.card.Card;
+import domain.card.Hand;
 import domain.participant.Dealer;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
