+++ src/test/java/service/BlackjackServiceTest.java
@@ -6,8 +6,8 @@ import constant.PlayerAction;
 import constant.PolicyConstant;
 import constant.Rank;
 import constant.Suit;
-import domain.Card;
 import domain.bet.Money;
+import domain.card.Card;
 import domain.participant.Names;
 import domain.participant.Participant;
 import domain.participant.Players;
